chore: rename icons.styles to iconsSizes
