+++ src/custom/CatalogFilter/CatalogFilter.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import VisibilityIcon from '@mui/icons-material/Visibility';
 import VisibilityOffIcon from '@mui/icons-material/VisibilityOff';
 import Button from '@mui/material/Button';
-import { iconMedium } from '../../constants/icons.styles';
+import { iconMedium } from '../../constants/iconsSizes';
 
 export interface CatalogFilterProps {
   catalogVisibility: boolean;
